Implement a lazily matching repeating parsers
This is somewhat tricky in the current form, where there is no sense of
backtracking.

The lookahead assertion takes a normal parser (of which the result is
returned) and another parser that should (or should not) match directly
after it.

For the "trivial" email address parsing example, also use the new
end-of-string assertition to only prevent matching TLDs at the very end.